refactor(favorites): refactor FavoritesStore with Zustand
