+++ src/screens/User/User.tsx
@@ -15,7 +15,6 @@ import AccountCircle from '../../icons/AccountCircle';
 import Favorite from '../../icons/Favorite';
 import BalanceWallet from '../../icons/BalanceWallet';
 import Exit from '../../icons/Exit';
-import { useFavorites } from '../../stores/FavoritesStore';
 import { useAccountStore } from '../../stores/AccountStore';
 import LoadingOverlay from '../../components/LoadingOverlay/LoadingOverlay';
 import ConfirmationDialog from '../../components/ConfirmationDialog/ConfirmationDialog';
@@ -25,6 +24,7 @@ import styles from './User.module.scss';
 
 import type { PlaylistItem } from '#types/playlist';
 import { logout } from '#src/stores/AccountController';
+import { clear as clearFavorites } from '#src/stores/FavoritesController';
 
 const User = (): JSX.Element => {
   const accessModel = useConfigStore((s) => s.accessModel);
@@ -37,7 +37,6 @@ const User = (): JSX.Element => {
   const { user: customer, subscription, transactions, activePayment, loading } = useAccountStore();
 
   const updateBlurImage = useBlurImageUpdater();
-  const { clearList: clearFavorites } = useFavorites();
 
   const onCardClick = (playlistItem: PlaylistItem) => history.push(cardUrl(playlistItem));
   const onCardHover = (playlistItem: PlaylistItem) => updateBlurImage(playlistItem.image);
